Moved build_matrices to helpers
